google-modules: bcm4398: Make DPC thread RT friendly
Now that dhd_sched_dpc() can be called from hard IRQ context, we need to
get rid of the spin locks that are added to the hard IRQ path when
switching away from tasklets in order to make this safe on PREEMPT_RT. The
wake lock isn't necessary and the other lock can be replaced by an atomic,
so make those changes in order to eliminate the spin locks.

The wake lock isn't necessary because the DPC kthread isn't flagged as
freezable, so it cannot freeze until it enters an _interruptible_ sleep.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>