KernelSU: Update ksu_handle_stat for 6.1
Our ksu_handle_stat hook will be updated for compatibility with 6.1,
and the KernelSU submodule needs to be updated accordingly or else
function declarations will be mismatched and shell will fail to
acquire root.

It appears as though they are planning to remove the
version-specific code from ksu_handle_stat in the future, but for
our purposes, it is necessary. Thus, split the function into
version-dependent variants and rewrite for explicit compatibility
with 6.1.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>