susfs4ksu: Tentatively introduce SUSFS_SUS_PATH [2/2]
Subsequent to a0aa01b ("KernelSU: Tentatively introduce SUSFS_SUS_PATH [1/2]")

We're importing this out of necessity, as it is currently the only
way to prevent detection of KSU's modules.img loop device.

Isolate this feature in a dedicated commit in hopes that this is a
temporary stop gap until a cleaner solution can be devised.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>