susfs4ksu: Tentatively restore SUSFS_SUS_PATH [2/2]
Subsequent to bb72f16 ("KernelSU: Tentatively restore SUSFS_SUS_PATH [1/2]")

We're bringing this back out of necessity, as it is currently the
only way to prevent detection of KSU's modules.img loop device.

Isolate this feature in a dedicated commit with hopes that this is
a temporary stop gap until a cleaner solution can be devised.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>