Import minimal subset of SUSFS v1.5.12 [2/2]
In a perfect world I wouldn't allow this sort of spaghetti code
anywhere near my kernel. Unfortunately, we live in a world which
has become increasingly hostile towards those of us who would
commit the grave sin of unlocking their bootloaders.

Whether I like it or not SUSFS is the most robust root-hiding
framework available, and while I can criticize it as a matter of
principle, I certainly cannot bring anything more effective to the
table.

What I can do is excise much of the feature bloat that SUSFS has
accumulated and pull in only the minimum functionality that is
required to achieve our base goal of effectively hiding root on
stock software. In this way, I can keep the hackery to an absolute
minimum which is a compromise that I'm willing to make for the
greater good.

[@0ctobot: Import based on:
https://gitlab.com/simonpunk/susfs4ksu/-/commit/f16560ce8263fbfa9b2f259e9531f72d6fda4e3f]
Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>