KernelSU: Spoof LKM mode flag to suppress manager warning
Starting from manager version 30000+, a warning card is displayed when
KernelSU is not detected as running in LKM mode. This detection is based
on bit 0 of the flags field returned by KSU_IOCTL_GET_INFO.

The original implementation only sets this bit when compiled with MODULE
defined meaning that it remains unset for direct kernel integration,
this triggers an unnecessary and frankly irrelevant warning as we are
not truly operating in either GKI or LKM mode.

Thus we will set the LKM bit at compile time to suppress this warning.
The flag is purely informational and used only by the manager for UI
display and debug logging, this change has no functional impact on
KernelSU operation.

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>