KernelSU: Conditionally hardcode KSU_VERSION for clarity
By default, KernelSU is configured to iterate versioning based on
number of revisions.

This becomes misleading in our case as making a change to the
submodule for compatibility with this kernel results in KSU_VERSION
incrementing from 32179 > 32180. This implies that we are running a
version of KSU that we actually aren't.

Hardcode KSU_VERSION to reflect the current release, and retain
default versioning logic if !CONFIG_LOCALVERSION="-NeutrinoKernel".

Signed-off-by: Adam W. Willis <return.of.octobot@gmail.com>